# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TODOリストアプリのRESTfulAPI（Go実装）",
        "title": "TodoList API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/tasks": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "ユーザのタスクを複数件取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスク取得",
                "parameters": [
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "description": "タスクステータス 0: 未着手 1: 完了",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Task"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "ユーザのタスクを追加する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスク追加",
                "parameters": [
                    {
                        "description": "Payload Description",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                }
            }
        },
        "/v1/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "ユーザのタスクを1件取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスク取得（1件）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "タスクID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "ユーザのタスクを更新する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスク更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "タスクID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload Description",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "ユーザのタスクを削除する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "タスク削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "タスクID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Task": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "タスク説明文",
                    "type": "string"
                },
                "id": {
                    "description": "タスクID（自動で生成されるUUID）",
                    "type": "string"
                },
                "status": {
                    "description": "タスクのステータス（0: 未着手, 1: 完了）",
                    "type": "integer"
                },
                "title": {
                    "description": "タスクタイトル",
                    "type": "string"
                },
                "user_id": {
                    "description": "ユーザID",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}